refactor: rename wallet asset component
